+++ src/Support/FileNamer/FileNamer.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Spatie\MediaLibrary\Support\FileNamer;
+
+use Spatie\MediaLibrary\Conversions\Conversion;
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
+
+abstract class FileNamer
+{
+    abstract public function conversionFileName(string $fileName, Conversion $conversion): string;
+
+    abstract public function responsiveFileName(string $fileName): string;
+
+    public function temporaryFileName(Media $media, string $extension): string
+    {
+        return "{$this->responsiveFileName($media->file_name)}.{$extension}";
+    }
+
+    public function extensionFromBaseImage(string $baseImage): string
+    {
+        return pathinfo($baseImage, PATHINFO_EXTENSION);
+    }
+}